+++ src/modules/ssh/mod.rs
@@ -33,6 +33,7 @@ pub mod path_lookup;
 pub mod path_roles;
 pub mod path_sign;
 pub mod policy;
+pub mod ssh_ca_audit_store;
 #[cfg(feature = "ssh_pqc")]
 pub mod pqc;